More google pricing

+++ src/com/xrtb/exchanges/google/GoogleWinObject.java
@@ -90,7 +90,7 @@ public class GoogleWinObject extends WinObject {
 	    
 	    DataInputStream dis = new DataInputStream( new ByteArrayInputStream(plaintext));
 	    
-	    value = dis.readDouble();
+	    value = dis.readLong();
 	    return value;
 	}
